requirements.txt
ruff
pre-commit
google-genai
python-dotenv
tqdm
